requirements.txt

openai>=1.61.0
agno>=2.4.0
python-dotenv==1.0.1
schedule==1.2.2
supabase==2.12.0
pillow==11.1.0
pydantic==2.10.1
pyyaml==6.0.1
playwright>=1.40.0
pytest>=8.0.0
pytest-mock>=3.12.0
psycopg-binary>=3.1.18
pgvector>=0.2.5
sqlalchemy>=2.0.28
feedparser>=6.0.10
nest_asyncio>=1.6.0
requests>=2.31.0
tweepy>=4.15.0
colorama>=0.4.6
playwright-stealth>=1.0.6
